requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
drf-yasg
django-filter
django-cors-headers
gunicorn
whitenoise
dj-database-url
